+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * InvalidChunk exception.
 * Support for VBRI headers in MP3.
+* ``LAMEEncodingFlags``.
+* ``ID3v2Flags``.
+* ABR presets to ``LAMEPreset`` enum.
 
 ### Changed